Implement showException as request parameter
The original implementation by Matthias Pigulla contained an
error in PreviewErrorController, where showException was set as a
request attribute instead of a query parameter. This was later
disimproved by Fabien Potencier by changing it from a request parameter
to an attribute in ExceptionController:showAction as well, which made
things consistent but unusable. The phpdoc of showAction() states the
original use case, which is also the only one I can think of that makes
sense: using showException as a request parameter. This commit restores
the originally intended behavior and changes/fixes the tests
accordingly.